requirements.txt
bs4
requests
telebot
pyTelegramBotAPI
python-dotenv
